Move all API work to a single container component. Extract utility functions.
